moved common.specs from e2e to mobile
