+++ tests/mobile/common/common.specs.js
@@ -1,4 +1,3 @@
-const helper = require('../../helper.js');
 const commonElements = require('../../page-objects/common/common.po.js');
 const utils = require('../../utils');
 
@@ -8,38 +7,34 @@ describe('Navigation tests : ', () => {
   it('should open Messages tab', () => {
     commonElements.goToMessages();
     expect(commonElements.isAt('message-list'));
+    expect(browser.getCurrentUrl()).toMatch(utils.getBaseUrl() + 'messages/');
   });
 
   it('should open tasks tab', () => {
     commonElements.goToTasks();
     expect(commonElements.isAt('task-list'));
+    expect(browser.getCurrentUrl()).toEqual(utils.getBaseUrl() + 'tasks/');
   });
 
   it('should open Reports or History tab', () => {
     commonElements.goToReports();
-    helper.handleUpdateModal();
     expect(commonElements.isAt('reports-list'));
   });
 
   it('should open Contacts or Peoples tab', () => {
     commonElements.goToPeople();
     expect(commonElements.isAt('contacts-list'));
+    expect(browser.getCurrentUrl()).toEqual(utils.getBaseUrl() + 'contacts/');
   });
 
   it('should open Analytics tab', () => {
     commonElements.goToAnalytics();
-    expect(element(by.css('div.targets')).isPresent());
+    expect(browser.getCurrentUrl()).toEqual(utils.getBaseUrl() + 'analytics');
   });
 
-  it('should open About', () => {
+  it('should open Configuration wizard', () => {
     commonElements.openMenu();
-    commonElements.checkAbout();
-  });
-
-  it('should open Report bug', () => {
-    helper.handleUpdateModal();
-    commonElements.openMenu();
-    commonElements.checkReportBug();
+    commonElements.checkConfigurationWizard();
   });
 
   it('should open Guided tour', () => {
@@ -47,8 +42,25 @@ describe('Navigation tests : ', () => {
     commonElements.checkGuidedTour();
   });
 
-  xit('should open User settings', () => {
+  it('should open About', () => {
+    commonElements.openMenu();
+    commonElements.checkAbout();
+  });
+
+  it('should open User settings', () => {
     commonElements.openMenu();
     commonElements.checkUserSettings();
+  });
+
+  it('should open Report bug', () => {
+    commonElements.openMenu();
+    commonElements.checkReportBug();
+  });
+
+  it('should open Configuration app', () => {
+    commonElements.goToConfiguration();
+    const display = element(by.css('[ui-sref="display.date-time"]'));
+    expect(display.isPresent()).toBeTruthy();
+    browser.get(utils.getBaseUrl() + 'messages/');
   });
 });
